work-connect-backend/workconnect/requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
gunicorn==20.1.0
mysqlclient==2.2.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
